Corrected face center

+++ app/src/main/java/com/firebase/hackweek/tank18thscale/PID.kt
@@ -6,25 +6,25 @@ class PID(private val kP : Float = 1f, private val kI: Float = 0f, private val k
 
     private var currentTime = LocalDateTime.now()
     private var previousTime = currentTime
-    private var previousError : Int = 0
-    private var cP : Long = 0
-    private var cI : Long = 0
-    private var cD : Long = 0
+    private var previousError : Double = 0.0
+    private var cP : Double = 0.0
+    private var cI : Double = 0.0
+    private var cD : Double = 0.0
 
-    fun update(error : Int, sleep : Long = 100) : Float {
+    fun update(error : Double, sleep : Long = 100) : Double {
         //Thread.sleep(sleep)
         currentTime = LocalDateTime.now()
         val deltaTime = Duration.between(currentTime, previousTime).toMillis()
 
         val deltaError = error - previousError
 
-        cP = error.toLong()
+        cP = error
         cI += error * deltaTime
 
         if (deltaTime > 0 ) {
             cD = deltaError  / deltaTime
         } else {
-            cD = 0
+            cD = 0.0
         }
 
         previousTime = currentTime
